little fixes.
posting the formatted from yalm dump for citation file,s as they likely more better.

releaser has some mods to make it easiewr to debug, and a ton of little fixes

moved file compression to before any requests. on hope the delay was causing crash. alas not. but not worth reverting