1) Update storing procedure of shape to avoid later on sorting procedure.
The keys are used as indexes to preserve the correct order.

The first shape dimension is the third packet received for each Tensor.
So based on this the allocation of dimensions to the array positions is
done.

In the Tensor value reception:
The start the indexing form zero, every time from the received
key the pre-vertex key is removed.